@@ -1,13 +1,21 @@
-import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
+import { RoutingModule } from './router/routing.module';
+import { CartIconModule } from './cart/cart-icon/cart-icon.module';
+import { MatToolbarModule } from '@angular/material/toolbar';
 import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
   declarations: [AppComponent],
-  imports: [BrowserModule, HttpClientModule],
-  providers: [],
+  imports: [
+    BrowserAnimationsModule,
+    HttpClientModule,
+    RoutingModule,
+    CartIconModule,
+    MatToolbarModule
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule {}
