@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 
 import { AppController } from './app.controller';
-import { AppService } from './app.service';
+import { ProductService } from './product/product.service';
 
 describe('AppController', () => {
   let app: TestingModule;
@@ -9,14 +9,14 @@ describe('AppController', () => {
   beforeAll(async () => {
     app = await Test.createTestingModule({
       controllers: [AppController],
-      providers: [AppService]
+      providers: [ProductService]
     }).compile();
   });
 
-  describe('getData', () => {
-    it('should return "Welcome to api!"', () => {
-      const appController = app.get<AppController>(AppController);
-      expect(appController.getData()).toEqual({ message: 'Welcome to api!' });
-    });
-  });
+  // describe('getData', () => {
+  //   it('should return "Welcome to api!"', () => {
+  //     const appController = app.get<AppController>(AppController);
+  //     expect(appController.getData()).toEqual({ message: 'Welcome to api!' });
+  //   });
+  // });
 });
