@@ -1,11 +1,13 @@
 import { Module } from '@nestjs/common';
 
 import { AppController } from './app.controller';
-import { AppService } from './app.service';
+import { ProductService } from './product/product.service';
+import { CartService } from './cart/cart.service';
+import { RatingService } from './rating/rating.service';
 
 @Module({
   imports: [],
   controllers: [AppController],
-  providers: [AppService]
+  providers: [ProductService, CartService, RatingService]
 })
 export class AppModule {}
