@@ -1,21 +1,21 @@
 import { Test } from '@nestjs/testing';
 
-import { AppService } from './app.service';
+import { ProductService } from './product/product.service';
 
 describe('AppService', () => {
-  let service: AppService;
+  let service: ProductService;
 
   beforeAll(async () => {
     const app = await Test.createTestingModule({
-      providers: [AppService]
+      providers: [ProductService]
     }).compile();
 
-    service = app.get<AppService>(AppService);
+    service = app.get<ProductService>(ProductService);
   });
 
-  describe('getData', () => {
-    it('should return "Welcome to api!"', () => {
-      expect(service.getData()).toEqual({ message: 'Welcome to api!' });
-    });
-  });
+  // describe('getData', () => {
+  //   it('should return "Welcome to api!"', () => {
+  //     expect(service.getData()).toEqual({ message: 'Welcome to api!' });
+  //   });
+  // });
 });
