@@ -1,13 +1,8 @@
 import { Component } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Message } from '@ngrx-nx-workshop/api-interfaces';
 
 @Component({
   selector: 'ngrx-nx-workshop-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
-export class AppComponent {
-  hello$ = this.http.get<Message>('/api/hello');
-  constructor(private http: HttpClient) {}
-}
+export class AppComponent {}
